WIP: updating tests
