+++ test/unit/consumer.test-todo.ts
@@ -94,453 +94,9 @@ describe("Consumer", () => {
     sandbox.restore();
   });
 
-  describe("Options Validation", () => {
-    it("requires an accountId to be set", () => {
-      assert.throws(() => new Consumer({}), {
-        message: "Missing consumer option [ accountId ].",
-      });
-    });
-
-    it("requires a queueId to be set", () => {
-      assert.throws(() => new Consumer({ accountId: ACCOUNT_ID }), {
-        message: "Missing consumer option [ queueId ].",
-      });
-    });
-
-    it("require a handleMessage or handleMessageBatch function to be set", () => {
-      assert.throws(
-        () => new Consumer({ accountId: ACCOUNT_ID, queueId: QUEUE_ID }),
-        {
-          message:
-            "Missing consumer option [ handleMessage or handleMessageBatch ].",
-        },
-      );
-    });
-
-    it("requires batchSize to be no greater than 100", () => {
-      assert.throws(
-        () =>
-          new Consumer({
-            accountId: ACCOUNT_ID,
-            queueId: QUEUE_ID,
-            handleMessage,
-            batchSize: 101,
-          }),
-        {
-          message: "batchSize must be between 1 and 100",
-        },
-      );
-    });
-
-    it("requires batchSize to be more than or equal to 1", () => {
-      assert.throws(
-        () =>
-          new Consumer({
-            accountId: ACCOUNT_ID,
-            queueId: QUEUE_ID,
-            handleMessage,
-            batchSize: -1,
-          }),
-        {
-          message: "batchSize must be between 1 and 100",
-        },
-      );
-    });
-
-    it("requires visibilityTimeoutMs to be less than 43200000", () => {
-      assert.throws(
-        () =>
-          new Consumer({
-            accountId: ACCOUNT_ID,
-            queueId: QUEUE_ID,
-            handleMessage,
-            visibilityTimeoutMs: 43200001,
-          }),
-        {
-          message: "visibilityTimeoutMs must be less than 43200000",
-        },
-      );
-    });
-
-    it("requires retryMessageDelay to be less than 42300", () => {
-      assert.throws(
-        () =>
-          new Consumer({
-            accountId: ACCOUNT_ID,
-            queueId: QUEUE_ID,
-            handleMessage,
-            retryMessageDelay: 42301,
-          }),
-        {
-          message: "retryMessageDelay must be less than 42300",
-        },
-      );
-    });
-  });
-
-  describe(".create", () => {
-    it("creates a new Consumer instance", () => {
-      consumer = Consumer.create({
-        accountId: ACCOUNT_ID,
-        queueId: QUEUE_ID,
-        handleMessage,
-      });
-
-      assert(consumer instanceof Consumer);
-    });
-  });
 
   describe(".start", () => {
     // TODO: Implement tests for the .start method
-    it("fires an event when the consumer starts", () => {
-      const handleStart = sandbox.stub().returns(null);
-      consumer.on("start", handleStart);
-
-      consumer.start();
-      consumer.stop();
-
-      sandbox.assert.calledOnce(handleStart);
-    });
-
-    it("calls the handleMessage function when a message is received", async () => {
-      consumer.start();
-      await pEvent(consumer, "message_processed");
-      consumer.stop();
-
-      sandbox.assert.calledWith(handleMessage, pullResponse.result.messages[0]);
-    });
-
-    it("calls the preReceiveMessageCallback and postReceiveMessageCallback function before receiving a message", async () => {
-      const preReceiveMessageCallbackStub = sandbox.stub().resolves(null);
-      const postReceiveMessageCallbackStub = sandbox.stub().resolves(null);
-
-      consumer = new Consumer({
-        accountId: ACCOUNT_ID,
-        queueId: QUEUE_ID,
-        handleMessage,
-        preReceiveMessageCallback: preReceiveMessageCallbackStub,
-        postReceiveMessageCallback: postReceiveMessageCallbackStub,
-      });
-
-      consumer.start();
-      await pEvent(consumer, "message_processed");
-      consumer.stop();
-
-      sandbox.assert.calledOnce(preReceiveMessageCallbackStub);
-      sandbox.assert.calledOnce(postReceiveMessageCallbackStub);
-    });
-
-    it("deletes the message when the handleMessage function is called", async () => {
-      handleMessage.resolves();
-
-      consumer.start();
-      await pEvent(consumer, "message_processed");
-      consumer.stop();
-
-      sandbox.assert.calledWithMatch(fetch, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          authorization: `Bearer ${QUEUES_API_TOKEN}`,
-        },
-        body: JSON.stringify({ ackIds: ["123"] }),
-      });
-    });
-
-    it("does not delete the message if shouldDeleteMessages is false", async () => {
-      consumer = new Consumer({
-        accountId: ACCOUNT_ID,
-        queueId: QUEUE_ID,
-        handleMessage,
-        shouldDeleteMessages: false,
-      });
-
-      handleMessage.resolves();
-
-      consumer.start();
-      await pEvent(consumer, "message_processed");
-      consumer.stop();
-
-      sandbox.assert.neverCalledWithMatch(fetch, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          authorization: `Bearer ${QUEUES_API_TOKEN}`,
-        },
-        body: JSON.stringify({ ackIds: ["123"] }),
-      });
-    });
-
-    it("doesn't delete the message when a processing error is reported", async () => {
-      handleMessage.rejects(new Error("Processing error"));
-
-      consumer.start();
-      await pEvent(consumer, "processing_error");
-      consumer.stop();
-
-      sandbox.assert.neverCalledWithMatch(fetch, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          authorization: `Bearer ${QUEUES_API_TOKEN}`,
-        },
-        body: JSON.stringify({ ackIds: ["123"] }),
-      });
-    });
-
-    it("consumes another message once one is processed", async () => {
-      handleMessage.resolves();
-
-      consumer.start();
-      await clock.runToLastAsync();
-      consumer.stop();
-
-      sandbox.assert.calledTwice(handleMessage);
-    });
-
-    it("doesn't consume more messages when called multiple times", () => {
-      fetch
-        .withArgs(PULL_MESSAGES_ENDPOINT, fetchArgs)
-        .resolves(new Promise((res) => setTimeout(res, 100)));
-      consumer.start();
-      consumer.start();
-      consumer.start();
-      consumer.start();
-      consumer.start();
-      consumer.stop();
-
-      fetch.calledOnceWith(PULL_MESSAGES_ENDPOINT, fetchArgs);
-    });
-
-    it("doesn't consume more messages when called multiple times after stopped", () => {
-      fetch
-        .withArgs(PULL_MESSAGES_ENDPOINT, fetchArgs)
-        .resolves(new Promise((res) => setTimeout(res, 100)));
-      consumer.start();
-      consumer.stop();
-
-      consumer.start();
-      consumer.start();
-      consumer.start();
-      consumer.start();
-
-      fetch.calledOnceWith(PULL_MESSAGES_ENDPOINT, fetchArgs);
-    });
-
-    it("consumes multiple messages when the batchSize is greater than 1", async () => {
-      fetch.withArgs(PULL_MESSAGES_ENDPOINT, fetchArgs).resolves({
-        ...pullResponse,
-        result: {
-          messages: [
-            ...pullResponse.result.messages,
-            {
-              body: "body-2",
-              id: "124",
-              timestamp_ms: 1234567890,
-              attempts: 1,
-              lease_id: "lease-id-2",
-              metadata: {
-                "CF-sourceMessageSource": "test",
-                "CF-Content-Type": "text",
-              },
-            },
-            {
-              body: "body-3",
-              id: "125",
-              timestamp_ms: 1234567890,
-              attempts: 1,
-              lease_id: "lease-id-3",
-              metadata: {
-                "CF-sourceMessageSource": "test",
-                "CF-Content-Type": "text",
-              },
-            },
-          ],
-        },
-      });
-
-      consumer = new Consumer({
-        accountId: ACCOUNT_ID,
-        queueId: QUEUE_ID,
-        handleMessage,
-        batchSize: 3,
-      });
-
-      consumer.start();
-      await pEvent(consumer, "message_received");
-      consumer.stop();
-
-      sandbox.assert.callCount(handleMessage, 3);
-      sandbox.assert.calledWithMatch(
-        fetch.firstCall,
-        PULL_MESSAGES_ENDPOINT,
-        fetchArgs,
-      );
-      sandbox.assert.match(
-        fetch.firstCall.args[0].input,
-        sinon.match({
-          path: "messages/pull",
-          method: "POST",
-          body: {
-            batch_size: 3,
-            visibility_timeout_ms: 30000,
-          },
-          accountId: ACCOUNT_ID,
-          queueId: QUEUE_ID,
-        }),
-      );
-    });
-
-    it("fires an emptyQueue event when all messages have been consumed", async () => {
-      fetch.withArgs(PULL_MESSAGES_ENDPOINT, fetchArgs).resolves({
-        ...pullResponse,
-        result: { messages: [] },
-      });
-
-      consumer.start();
-      await pEvent(consumer, "empty");
-      consumer.stop();
-    });
-
-    it("terminate message visibility timeout on processing error", async () => {
-      handleMessage.rejects(new Error("Processing error"));
-
-      consumer.terminateVisibilityTimeout = true;
-
-      consumer.start();
-      await pEvent(consumer, "processing_error");
-      consumer.stop();
-
-      sandbox.assert.calledWith(fetch.secondCall, mockChangeMessageVisibility);
-      sandbox.assert.match(
-        sqs.send.secondCall.args[0].input,
-        sinon.match({
-          accountId: ACCOUNT_ID,
-          ReceiptHandle: "receipt-handle",
-          VisibilityTimeout: 0,
-        }),
-      );
-    });
-
-    it("does not terminate visibility timeout when `terminateVisibilityTimeout` option is false", async () => {
-      handleMessage.rejects(new Error("Processing error"));
-      consumer.terminateVisibilityTimeout = false;
-
-      consumer.start();
-      await pEvent(consumer, "processing_error");
-      consumer.stop();
-
-      sqs.send.neverCalledWith(mockChangeMessageVisibility);
-    });
-
-    it("calls the handleMessagesBatch function when a batch of messages is received", async () => {
-      consumer = new Consumer({
-        accountId: ACCOUNT_ID,
-        queueId: QUEUE_ID,
-        handleMessageBatch,
-        batchSize: 2,
-      });
-
-      consumer.start();
-      await pEvent(consumer, "response_processed");
-      consumer.stop();
-
-      sandbox.assert.callCount(handleMessageBatch, 1);
-    });
-
-    it("handles unexpected exceptions thrown by the handler batch function", async () => {
-      consumer = new Consumer({
-        accountId: ACCOUNT_ID,
-        queueId: QUEUE_ID,
-        handleMessageBatch: () => {
-          throw new Error("unexpected parsing error");
-        },
-        batchSize: 2,
-      });
-
-      consumer.start();
-      const err: any = await pEvent(consumer, "error");
-      consumer.stop();
-
-      assert.ok(err);
-      assert.equal(
-        err.message,
-        "Unexpected message handler failure: unexpected parsing error",
-      );
-    });
-
-    it("handles non-standard objects thrown by the handler batch function", async () => {
-      class CustomError {
-        private _message: string;
-
-        constructor(message) {
-          this._message = message;
-        }
-
-        get message() {
-          return this._message;
-        }
-      }
-
-      consumer = new Consumer({
-        accountId: ACCOUNT_ID,
-        queueId: QUEUE_ID,
-        handleMessageBatch: () => {
-          throw new CustomError("unexpected parsing error");
-        },
-        batchSize: 2,
-      });
-
-      consumer.start();
-      const err: any = await pEvent(consumer, "error");
-      consumer.stop();
-
-      assert.ok(err);
-      assert.equal(err.message, "unexpected parsing error");
-    });
-
-    it("handles non-standard exceptions thrown by the handler batch function", async () => {
-      const customError = new Error();
-      Object.defineProperty(customError, "message", {
-        get: () => "unexpected parsing error",
-      });
-
-      consumer = new Consumer({
-        accountId: ACCOUNT_ID,
-        queueId: QUEUE_ID,
-        handleMessageBatch: () => {
-          throw customError;
-        },
-        batchSize: 2,
-      });
-
-      consumer.start();
-      const err: any = await pEvent(consumer, "error");
-      consumer.stop();
-
-      assert.ok(err);
-      assert.equal(
-        err.message,
-        "Unexpected message handler failure: unexpected parsing error",
-      );
-    });
-
-    it("prefers handleMessagesBatch over handleMessage when both are set", async () => {
-      consumer = new Consumer({
-        accountId: ACCOUNT_ID,
-        queueId: QUEUE_ID,
-        handleMessageBatch,
-        handleMessage,
-        batchSize: 2,
-      });
-
-      consumer.start();
-      await pEvent(consumer, "response_processed");
-      consumer.stop();
-
-      sandbox.assert.callCount(handleMessageBatch, 1);
-      sandbox.assert.callCount(handleMessage, 0);
-    });
 
     it("ack the message if handleMessage returns void", async () => {
       consumer = new Consumer({
